Task 1 - Ising model
I believe everything here is correct. If something is off feel free to let me know.

The calculate_agreement function passes all the tests in ising_test
The outputs with various values of alpha and external {external=0 α=0.01, external=0 α=10, external=0.1 α=0.001, external=0.1 α=0.5} seem appropriate when considering the images in the blackboard document.