Update Task 1 attempt.py
Same as before just rewrote some of the variable names to make them more "conventional"

I also realised that the program needs to be able to accept values for external and alpha when running the program from terminal. Will do this another time I don't actually remember how to do that lol.